# policy: use netip.Addr when constructing CIDR rules