--- a/pkg/ip/ip_test.go
+++ b/pkg/ip/ip_test.go
@@ -774,6 +774,119 @@ func (s *IPTestSuite) BenchmarkKeepUniqueIPs(c *C) {
 	}
 }
 
+func TestKeepUniqueAddrs(t *testing.T) {
+	for _, tc := range []struct {
+		name  string
+		addrs []netip.Addr
+		want  []netip.Addr
+	}{
+		{
+			name:  "nil slice",
+			addrs: nil,
+			want:  nil,
+		},
+		{
+			name:  "empty slice",
+			addrs: []netip.Addr{},
+			want:  []netip.Addr{},
+		},
+		{
+			name:  "one element slice",
+			addrs: []netip.Addr{netip.MustParseAddr("1.1.1.1")},
+			want:  []netip.Addr{netip.MustParseAddr("1.1.1.1")},
+		},
+		{
+			name: "IPv4 all duplicates",
+			addrs: []netip.Addr{
+				netip.MustParseAddr("1.1.1.1"),
+				netip.MustParseAddr("1.1.1.1"),
+				netip.MustParseAddr("1.1.1.1"),
+			},
+			want: []netip.Addr{
+				netip.MustParseAddr("1.1.1.1"),
+			},
+		},
+		{
+			name: "IPv4 all unique",
+			addrs: []netip.Addr{
+				netip.MustParseAddr("1.1.1.1"),
+				netip.MustParseAddr("2.2.2.2"),
+				netip.MustParseAddr("3.3.3.3"),
+			},
+			want: []netip.Addr{
+				netip.MustParseAddr("1.1.1.1"),
+				netip.MustParseAddr("2.2.2.2"),
+				netip.MustParseAddr("3.3.3.3"),
+			},
+		},
+		{
+			name: "IPv4 mixed",
+			addrs: []netip.Addr{
+				netip.MustParseAddr("3.3.3.3"),
+				netip.MustParseAddr("1.1.1.1"),
+				netip.MustParseAddr("2.2.2.2"),
+				netip.MustParseAddr("2.2.2.2"),
+			},
+			want: []netip.Addr{
+				netip.MustParseAddr("1.1.1.1"),
+				netip.MustParseAddr("2.2.2.2"),
+				netip.MustParseAddr("3.3.3.3"),
+			},
+		},
+		{
+			name: "IPv6 all duplicates",
+			addrs: []netip.Addr{
+				netip.MustParseAddr("f00d::1"),
+				netip.MustParseAddr("f00d::1"),
+				netip.MustParseAddr("f00d::1"),
+			},
+			want: []netip.Addr{
+				netip.MustParseAddr("f00d::1"),
+			},
+		},
+		{
+			name: "Mixed IPv4 & IPv6",
+			addrs: []netip.Addr{
+				netip.MustParseAddr("::1"),
+				netip.MustParseAddr("f00d::1"),
+				netip.MustParseAddr("1.1.1.1"),
+				netip.MustParseAddr("f00d::1"),
+			},
+			want: []netip.Addr{
+				netip.MustParseAddr("1.1.1.1"),
+				netip.MustParseAddr("::1"),
+				netip.MustParseAddr("f00d::1"),
+			},
+		},
+		{
+			name: "With IPv6-in-IPv6",
+			addrs: []netip.Addr{
+				netip.MustParseAddr("::ffff:0101:0101"),
+				netip.MustParseAddr("::ffff:1.1.1.1"),
+				netip.MustParseAddr("1.1.1.1"),
+			},
+			want: []netip.Addr{
+				netip.MustParseAddr("1.1.1.1"),
+				netip.MustParseAddr("::ffff:1.1.1.1"),
+			},
+		},
+	} {
+		t.Run(tc.name, func(t *testing.T) {
+			got := KeepUniqueAddrs(tc.addrs)
+			if len(tc.want) != len(got) {
+				t.Errorf("%s: KeepUniqueAddrs(%q): got %d unique addresses, want %d",
+					tc.name, tc.addrs, len(got), len(tc.want))
+			}
+			for i := range got {
+				if tc.want[i] != got[i] {
+					t.Errorf("%s: KeepUniqueAddrs(%q): mismatching address at index %d: got %v, want %v",
+						tc.name, tc.addrs, i, got[i], tc.want[i])
+				}
+			}
+		})
+	}
+}
+
 func (s *IPTestSuite) TestIPVersion(c *C) {
 	type args struct {
 		ip net.IP