--- a/pkg/policy/api/cidr.go
+++ b/pkg/policy/api/cidr.go
@@ -5,6 +5,7 @@ package api
 
 import (
 	"net"
+	"net/netip"
 	"strings"
 
 	"github.com/cilium/cilium/pkg/ip"
@@ -154,16 +155,17 @@ func ComputeResultantCIDRSet(cidrs CIDRRuleSlice) CIDRSlice {
 	return allResultantAllowedCIDRs
 }
 
-// IPsToCIDRRules generates CIDRRules for the IPs passed in./
+// addrsToCIDRRules generates CIDRRules for the IPs passed in.
 // This function will mark the rule to Generated true by default.
-func IPsToCIDRRules(ips []net.IP) (cidrRules []CIDRRule) {
-	for _, ip := range ips {
+func addrsToCIDRRules(addrs []netip.Addr) []CIDRRule {
+	cidrRules := make([]CIDRRule, 0, len(addrs))
+	for _, addr := range addrs {
 		rule := CIDRRule{ExceptCIDRs: make([]CIDR, 0)}
 		rule.Generated = true
-		if ip.To4() != nil {
-			rule.Cidr = CIDR(ip.String() + "/32")
+		if addr.Is4() {
+			rule.Cidr = CIDR(addr.String() + "/32")
 		} else {
-			rule.Cidr = CIDR(ip.String() + "/128")
+			rule.Cidr = CIDR(addr.String() + "/128")
 		}
 		cidrRules = append(cidrRules, rule)
 	}