--- a/pkg/policy/api/groups.go
+++ b/pkg/policy/api/groups.go
@@ -6,7 +6,7 @@ package api
 import (
 	"context"
 	"fmt"
-	"net"
+	"net/netip"
 	"sync"
 
 	"github.com/cilium/cilium/pkg/ip"
@@ -22,7 +22,7 @@ var (
 
 // GroupProviderFunc is a func that need to be register to be able to
 // register a new provider in the platform.
-type GroupProviderFunc func(context.Context, *ToGroups) ([]net.IP, error)
+type GroupProviderFunc func(context.Context, *ToGroups) ([]netip.Addr, error)
 
 // ToGroups structure to store all kinds of new integrations that needs a new
 // derivative policy.
@@ -47,8 +47,7 @@ func RegisterToGroupsProvider(providerName string, callback GroupProviderFunc) {
 // GetCidrSet will return the CIDRRule for the rule using the callbacks that
 // are register in the platform.
 func (group *ToGroups) GetCidrSet(ctx context.Context) ([]CIDRRule, error) {
-
-	var ips []net.IP
+	var addrs []netip.Addr
 	// Get per  provider CIDRSet
 	if group.AWS != nil {
 		callbackInterface, ok := providers.Load(AWSProvider)
@@ -59,15 +58,15 @@ func (group *ToGroups) GetCidrSet(ctx context.Context) ([]CIDRRule, error) {
 		if !ok {
 			return nil, fmt.Errorf("Provider callback for %s is not a valid instance", AWSProvider)
 		}
-		awsIPs, err := callback(ctx, group)
+		awsAddrs, err := callback(ctx, group)
 		if err != nil {
 			return nil, fmt.Errorf(
 				"Cannot retrieve data from %s provider: %s",
 				AWSProvider, err)
 		}
-		ips = append(ips, awsIPs...)
+		addrs = append(addrs, awsAddrs...)
 	}
 
-	resultIps := ip.KeepUniqueIPs(ips)
-	return IPsToCIDRRules(resultIps), nil
+	resultAddrs := ip.KeepUniqueAddrs(addrs)
+	return addrsToCIDRRules(resultAddrs), nil
 }