--- a/pkg/policy/api/groups_test.go
+++ b/pkg/policy/api/groups_test.go
@@ -8,7 +8,8 @@ package api
 import (
 	"context"
 	"fmt"
-	"net"
+	"net/netip"
+	"testing"
 
 	. "gopkg.in/check.v1"
 
@@ -31,16 +32,16 @@ func GetToGroupsRule() ToGroups {
 	}
 }
 func GetCallBackWithRule(ips ...string) GroupProviderFunc {
-	netIPs := []net.IP{}
+	netIPs := make([]netip.Addr, 0, len(ips))
 	for _, ip := range ips {
-		netIPs = append(netIPs, net.ParseIP(ip))
+		if addr, err := netip.ParseAddr(ip); err == nil {
+			netIPs = append(netIPs, addr)
+		}
 	}
 
-	cb := func(ctx context.Context, group *ToGroups) ([]net.IP, error) {
+	return func(ctx context.Context, group *ToGroups) ([]netip.Addr, error) {
 		return netIPs, nil
 	}
-
-	return cb
 }
 
 func (s *PolicyAPITestSuite) TestGetCIDRSetWithValidValue(c *C) {
@@ -83,9 +84,8 @@ func (s *PolicyAPITestSuite) TestGetCIDRSetWithUniqueCIDRRule(c *C) {
 }
 
 func (s *PolicyAPITestSuite) TestGetCIDRSetWithError(c *C) {
-
-	cb := func(ctx context.Context, group *ToGroups) ([]net.IP, error) {
-		return []net.IP{}, fmt.Errorf("Invalid credentials")
+	cb := func(ctx context.Context, group *ToGroups) ([]netip.Addr, error) {
+		return []netip.Addr{}, fmt.Errorf("Invalid credentials")
 	}
 	RegisterToGroupsProvider(AWSProvider, cb)
 	group := GetToGroupsRule()
@@ -101,4 +101,18 @@ func (s *PolicyAPITestSuite) TestWithoutProviderRegister(c *C) {
 	cidr, err := group.GetCidrSet(context.TODO())
 	c.Assert(cidr, IsNil)
 	c.Assert(err, NotNil)
+}
+
+func BenchmarkGetCIDRSet(b *testing.B) {
+	cb := GetCallBackWithRule("192.168.1.1", "192.168.10.10", "192.168.10.3")
+	RegisterToGroupsProvider(AWSProvider, cb)
+	group := GetToGroupsRule()
+	b.ReportAllocs()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_, err := group.GetCidrSet(context.TODO())
+		if err != nil {
+			b.Fatal(err)
+		}
+	}
 }