First commit adds a helper function, second commit converts CIDR rules to use netip.Addr internally. This should somewhat improve memory usage when using CIDR rules in policies.

Before/after benchmark:
    
    name          old time/op    new time/op    delta
    GetCIDRSet-8    2.58µs ± 6%    1.96µs ± 6%  -23.84%  (p=0.000 n=20+20)
    
    name          old alloc/op   new alloc/op   delta
    GetCIDRSet-8      608B ± 0%      416B ± 0%  -31.58%  (p=0.000 n=20+20)
    
    name          old allocs/op  new allocs/op  delta
    GetCIDRSet-8      13.0 ± 0%      11.0 ± 0%  -15.38%  (p=0.000 n=20+20)

See commits for details.