@tklauser 2022-09-21
> lgtm (no before/after benchmarks? slightly_smiling_face )

Good point. Added benchmarks to the last commit and the PR description.